chore: disable snapshotting feature by default
This patch ensures that balance snapshots based caching is disabled
by default. It can be enable to speed up queries, however any inconsistencies
in the cache itself causes a ripple effect that queries for the value
endpoints will fail indefinetly.

We will enable this in the forthcoming releases when the calculation
algorithm has been already verified to be stable.