fix: required and readonly fields be optional
If the a field in OpenAPI spec is marked as readonly and required at the
same time the Go code generator makes the field optional by setting its
type to pointer which is not the desired behaviour.

Fortunatelly it is possible to disable this behaviour in teh generator
configuration.

See: https://github.com/oapi-codegen/oapi-codegen/issues/778