feat: billing profiles (#1703)
This patch adds support for billing profiles as the first 
implemented API.

For testing purposes, a new OpenMeter Sandbox is added. The sandbox 
app gets installed by default if apps are enabled for easier testing.

The change also makes sure that the customeroverrides codebase 
works as expected, however we are going to implement the APIs at a 
later point.