feat: validation errors api (#1772)
Given that most of the invoice operations will be done in the background, 
we need to capture the validation errors. This patch does it in a way that is 
native to go's error handling: an error can be parsed into multiple validation
 errors, if an error is not a validationissue typed error or not wrapped into any 
validation issue modifier it is treated as a normal error.

The state machine honors the errors: if there are critical errors that's 
considered a failed state.

The validation errors are available on the invoice endpoint.