fix: watermill retry logic and error reporting
This patch ensures that the max retires is set, if we have a retry
timeout, as without the max retries watermill would never retry.

Given that in our use-case we also have DLQ, that would print an error
if an issue happens, we decrease the severity of a retry message to
warning.