feat: only syncronize lines that have changed
This allows us to properly maintain the updatedAt of the invoice base, and
decreases the amount of invoice update events.

The patch only fixes the sync, as capturing this in the adapter is a more
complex venture, also except of gathering invoices, if we edit an invoice
there will be an update due to the status change either ways.