fix: customeroverride overload
Gathering invoice handling needs a lock to prevent parallel executions
of the gathering invoice creation.

Previously we were using the billing_customeroverride table for this
however that results in a lot of empty rows there, which makes it hard
to understand the user intent behind the row.

This patch modifies this behavior by creating a dedicated lock table that
we can use for gathering invoice coordination if needed.