fix: progressive billing timing issues
This patch unifies the locking mechanism (previously we were using
customer and invoice locks), as most of progressive billing actions
require multiple invoice locks.

The patch also fixes the issue that if a syncronization would yield
an empty ubp line, then we threw an error.

The patch also makes sure that progressively billed split lines do
not prevent the deletion of the gathering invoice.