fix: phase iterator in advance/in arreas
If in advance and in arreas items are mixed on an invoice, the
current phase iterator does not generate the in advance item
for the next phase.

This results in the in advance and in arreas items to be included in
seperate invoices.

This patch fixes this by considering the invoice_at property of the
soon to be generated lines.