fix: only support qty=1 for flat fee lines
Given that product catalog drives invoicing (we are not used as standalone
invoicing engine) and flat price only supports qty=1 we are no longer
allowing flat fee lines with other qty on the invoice's root.

Detailed lines are also represented as flat fee lines, there we allow qty
to be anything positive.