fix: do not sync old invoices
With the recent changes the invoice at is not truncated anymore (which
is a good behavior).

Given previously it was truncated the sync logic tries to reconcile the
old already issued invoices. This is done by checking the total sum and
if there is any difference a warning is added.

Given the invoice at is not used for non-gathering invoices this is just
a waste of resources. This patch makes sure that we are not managing the
invoice at fields for lines that are not on gathering invoices.

Going forward we should drop the invoice_at from the lines on the
non-gathering invoices.