fix: improve watermill router shutdown behavior
This patch makes sure that if the router is shutting down (closing), we
are disabling the poision queue and do not ack the message to Kafka.

This is required as the Router's close almost immediately cancels the
message's context, thus the processing will immedietly stop.

We are also making sure that we treat these errors as warnings instead
of Errors, as due to the NAck (noop for Kafka) we will retry after restart.