chore: reduce worker noise
This patch logs the entitlement not found errors as warnings, given that
rolled back transactions also fire events.

Due to this we can expect that sometimes the entitlement is not found.

We still need to retry, as if the originating transaction is running for a
while, the entitlement might appear after the transaction is committed.