fix: customer delete
Use a different field to determine if the usage_attribution of the subject
has been deleted due to the customer being deleted or if the mapping has changend.

I would prefer this approach instead of checking for equality with the customer.deleted_at
field as it makes it more explicit why the deletion have happened.