fix: stripe app errors
Setting the same `stripe_customer_id` for multiple Subjects is allowed
as uniqueness is not enforced. This causes problems when it comes to
upserting Customer entities for Subjects as `stripe_customer_id` must be
uniqe for each Customer.

This change will surface these errors to users so they can decide how to
resolve this issues as we cannot decide on backend if the user assgined
the `stripe_customer_id` to Subject as metadata or it is used for one of
the Stripe integrations (Stripe Usage Reporter, Stripe App, etc).

In case of the former we encourage users to use the `metadata` field of
the Subject entity to store any metadata, otherwise users need to ensure
that `stripe_customer_id` are unique across Subject entities in order
to resolve this issue.