feat: add clickhouse retries (#3691)
This patch adds support for retring certian known-to-be retriable clickhouse errors:

- EOF: happens when node rotation occurs due to clickhouse-go not pinging the server
- Connection pool exhaused: worth retying with warnings being logged

Only select queries are retried.

This should be enabled for backend only, as the workers are capable of retries using Watermill.

We are not using the usual centakill/backoff here as that's an overkill compared to what we want to achieve (also that has inverse logic: some errors are permanent, not a few errors are retryable).